Fix script name issue
I accidentally renamed the old script for updating pansolid cnv data to "live_service_collate_cnv_data.R" so I have corrected this.
I've also moved the new collation script out of the dev folder and renamed it to make clear that this is the script for collating PanSolid CNV data after the go-live date for deletions and LOH detection.